Update addr_counter file and remove output enable function
At this point, because of the warnings messages Quartus II was giving me
about internal tri-state buses being impossible to synthesize, I decided
to create a new HDL component that is essentially a bus multiplexer.

I needed to remove the output enable function on the addr_counter, since this
signal will go directly to the bus mux instead.