Get rid of output enable signal
For the current design, the LED segment register is the only device
that is accessing the RAM data line. There is thus no need for tri-state
output for the LED segment register

In the future, another bus multiplexer might be implemented because of
the ROM module which will have to write to the RAM. However, an output
enable signal is still not necessary for the LED segment register block.