Update with various component improvements
Import single component shift register stage and
the 8th bit for the address decoder (the output enable signal for
the address counter).

This signal will be used for ensuring during the
write cycle of the RAM module, the column selector does not switch to the
column set by the position register.