Update symbol formatting (text size) and add PLL
When switching the display resolution of my screen, the text in the symbol
boxes of the previous full_design schematics were crossing the symbol box boundaries.

This update ensures that the text is contained to its boundary.

Add a PLL that takes in the 50 MHz onboard clock signal and converts it
to a 100 kHz clock for the different modules.