Fix bug - incorrect port ordering for FSM component
The last version of the full_design schematic was used for ModelSim simulation.
I was not able to recreate the waveforms obtained from the University Program
Waveform simulator and I was trying to figure out why. After careful inspection
of the schematic and observing the erroneous simulation behavior, I realized that the
COL_ON, ROW_ON, and DIG_ON signals were connected to the incorrect FSM inputs.

Also an update to the naming convention - bus signals are lower case and bit signals
are uppercase.