Add bus mux, clean up signal routing, and use PLL for inverse clock
This is the schematic where the new bus multiplexer is used instead of
the tri-state buses internal to the address counter and position register modules.

Also rearrange the port signals of various modules so that less signals
intersect with each other in the schematic.

Also modify the PLL so that 2 module clocks can be obtained that are 180 degrees
out of phase with each other (CLK and INV_CLK). Before, I thought that the only way to do this was
by using a NOT gate. However within the IP config interface, it was possible to add the
phase offset.