Add open-drain output primitives for the column selection signals
My initial thought was that by using open drain primitives, a much larger
current can be sinked by the associated pins. This large current will come
from all the LED segment lines for each row and for a particular column.

However, after discovering that the open-drain is essentially a tri-state buffer
with an inverted output enable signal, I am skeptical on whether or not the max
current limit is increased. Further investigation is needed.