Delete all the internal signals for simulation/debug purposes
This schematic gets rid of all of the internal FSM signals, the address lines,
encoder signals, etc. as outputs on the schematic. This is because
these outputs were only used during ModelSim simulation to verify
the correct functionality of the design.

What rests is only the inputs and outputs of the actual Sudoku design.

Another addition was a global RESET signal for the various components.
This is an inverted reset line because I am using a push button
that is onboard the MAX 10 FPGA dev board and is active low
(the button logic state is high when not pressed).

If further simulation is required, think to revert to the previous commit
and add the global reset line so that all the internal signals can be accessed.