Update full_design schematic with up-to-date HDL modules
Remove the version tags (_v01 for example) on all the HDL
symbol blocks in the full_design .bdf file.

Also cleaned up the signals on the schematic and some wiring/nets.

Also ran a compilation with the full_design schematic as the top-level
entity and no issues were found. Pending simulation.