Fix the MAX_CLOCK_CYCLE bug in the input_timeout HDL file
When running simulation, now the DELAY_DONE signal becomes active

Before, what was happening was that I was performing an integer math
operation of SECONDS --> MS or more specifically the TIMEOUT/1E+3

This operation in decimal math would have given 0.001 or 1 ms, but
in integer math this was turned into "0". This means
that the MAX_CLOCK_CYCLE would ALWAYS be 0.

When looking at the VHDL code, I see that I would increment the tick
variable immediately after the rising_edge call. Thus, the tick variable
could never = 0. That is why the DELAY_DONE signal is never activated