Test with external LED driver (ULN2003A)
Success with slow PLL clock (1.2 kHz). Due to signal switching,
my ampmeter was unable to obtain an accurate value for the current passing
through each column once that column is activated.

However, first visual results seem promising. Also noticed
during the tests that there that is clearly button bouncing even with the RC circuit.
Tomorrow I'll work on that and test how much DC current is passing when
the columns have all their segments lit.

Other changes - remove open_drain primitives and weak_pull_up assignments
Also change the address_decoder entity so that when a column is selected,
it is activated (to select column 1 for example, the output bus has to be 000000001)