Test hardware with 100 kHz PLL clock and observe luminosity
Since each column is briefly
turned on and cycled in the full Sudoku design,
the *average* current of each column
branch is significantly lower

This leads to lower luminositiy levels. For testing purposes this is
sufficient. However for the final design, a higher current gain driver
should be used so that more LED current can be sinked for illuminating
an entire column.

Branch will be merged in to master since the external LED driver
feature has been implemented with success.