Simulate design with inverted clocks for all blocks except for FSM
In an effort to debug the hardware and to obtain results similiar to simulation,
my first idea was to use inverted clocks not only for the registers,
but also for all components that the FSM controls.

My thinking was that this will remove the ambiguity between when the FSM control
signals arrive and when the components make a decision based on the states
of these control signals.

In theory, this should improve the low hold times and perhaps
fix the issues that are arriving when the hardware is tested at
a PLL clock frequency of 100 kHz.

As expected, ModelSim yields basically unchanged results: the inverted
clock shouldn't change the process of how the system writes to memory.