Update testbench and run simulation with software debouncing
ModelSim simulation was successful - just like before, the
appropiate column/row/digit is written to memory
even with the newly added software debouncing module

Although the debouncing function was not tested in simulation, this
can be done in the future by simply writing a test bench that
simulates the rapid bouncing of the keyboard signals.
However, since the module acts as a simple delay, this simulation isn't
necessary for verifying the correct functionality.