Test hardware with new keyboard design (internal pullups)
Hardware test was a success. The design works as before when
there were external pull-down resistors.

A little bit of troubleshooting was done during testing:

On the breadboard, I previously had the other keyboard signals connected
to ground. With the new keyboard design, the keyboard states are now
pulled high when the buttons are not activated. Thus, since I had
all the other keyboard signals low, my HDL code for the encoders
gave a "Don't care" output for the signals going to the registers.

Another issue - one of my jumper cables that went to the digit "1"
was not well connected. After changing out this cable I was able
to write "1" to the LED matrix