Fix issues with pin swapping on FPGA connectors
The nets weren't being registered because I didn't have a wire
connecting the resistors to the jumper header connectors.

Also, realized that I was incorrectly naming the header parts pins ->
since the FPGA headers are added with jumper wires, I can pin swap
all the pins on the PCB headers and these headers shouldn't have any names.

I added at the bottom of the schematic an explanation of this, hopefully it
makes sense. Essentially, as long as each header isn't exceeding the number of pins
available from the FPGA headers, we can swap all the pins around on the PCB and
afterwards just connect the corresponding jumper wires after pin configuration in Quartus
pin planner.