Quickly route 'key' nets
There is a slight issue with synchronization between the schematic and the PCB.

I think this is because I swapped pins on the PCB and then starting routing nets, but
didn't push changes to schematic after swapping. One net example is J-23 and J-24. If
I cross probe one of those nets the incorrect net is highlighted on the schematic
and the PCB. This might be manually resolved but I'm not sure of the best way to do that
(match nets? Generate custom ECO?)

Things left to route -> all the extra game options (and LEDs associated),
the KEY[9..0] bus, the LED column signals, power and ground,
and the DIG_ON, COL_ON, and ROW_ON.