Fix naming LED#ROW#, clean up connectors, add net names
Realized that every bus had the opposite designator for the bus, that is to say
that each entry to the bus should be LED0, LED1... etc. instead of ROW0, ROW1, etc.

Got rid of the non-generic headers and replaced them with simple 20x2 headers from
Altium Library. Also matched snap grid and aligned, so that attaching the components and
nets was much easier.

Finally, added descriptive net names for connectors between resistors. This is for
synchronizing the PCB and schematic (errors due to mismatch after pin swapping).