Synchronize schematic and PCB
I'm pretty sure all the synchronization is done, as I've run update schematic to
PCB and vice versa with no changes to report (except for Room definitions). Hopefully
all the nets are in the right place, which I'm fairly confident about since the schematic
doesn't show any issues with how the PCB is routed.